Make gpmd link a variable and make all vars upper
The link for google play music desktop player is long, and will most
likely change. It will be a good idea to store it as a variable
in order to modify it more easily, as well as improve how the code
looks to a small degree.

Not all of my variables were upper case. I have corrected this so that
the naming scheme is consistent.